formating

+++ jtracker/execution/worker.py
@@ -61,7 +61,8 @@ def download_file(local_path, url, logger):
         except Exception as e:
             if os.path.isfile(local_path):
                 os.remove(local_path)  # remove unfinished file
-            os.remove(local_path + '.__downloading__')  # remove flag
+            if os.path.isfile(local_path + '.__downloading__'):  # just in case
+                os.remove(local_path + '.__downloading__')  # remove flag
             raise Exception(e)
 
         # update the flag to indicate file is ready
@@ -181,38 +182,38 @@ class Worker(object):
             with open(os.path.join(self.task_dir, '_file_provision_err.txt'), 'a') as f:
                 f.write(file_provision_error)
         else:
-          command = self._task_command_builder()
-          self.logger.debug("Task command is: %s" % command)
+            command = self._task_command_builder()
+            self.logger.debug("Task command is: %s" % command)
 
-          for n in range(retry + 1):
-            success = True  # assume task complete
-            if n > 0:
-                pause = 100 * 2 ** n
-                self.logger.info('Task: %s failed, retry in %s seconds; job: %s' %
-                      (self.task.get('name'), pause, self.task.get('job.id')))
-                sleep(pause)  # pause before retrying
-                self.logger.info('No %s retry on task: %s; job: %s' %
-                      (n, self.task.get('name'), self.task.get('job.id')))
-            try:
-                p = subprocess.Popen([command], stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
-                stdout, stderr = p.communicate()
-            except Exception as e:
-                success = False
+            for n in range(retry + 1):
+                success = True  # assume task complete
+                if n > 0:
+                    pause = 100 * 2 ** n
+                    self.logger.info('Task: %s failed, retry in %s seconds; job: %s' %
+                          (self.task.get('name'), pause, self.task.get('job.id')))
+                    sleep(pause)  # pause before retrying
+                    self.logger.info('No %s retry on task: %s; job: %s' %
+                          (n, self.task.get('name'), self.task.get('job.id')))
+                try:
+                    p = subprocess.Popen([command], stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
+                    stdout, stderr = p.communicate()
+                except Exception as e:
+                    success = False
 
-            if p.returncode != 0 or success is False:
-                with open(os.path.join(self.task_dir, 'stdout.txt'), 'a') as o:
-                    o.write("Run no: %s, STDOUT at: %s\n" % (n+1, int(time())))
-                    o.write(stdout.decode("utf-8"))
-                with open(os.path.join(self.task_dir, 'stderr.txt'), 'a') as e:
-                    e.write("Run no: %s, STDERR at: %s\n" % (n+1, int(time())))
-                    e.write(stderr.decode("utf-8"))
-                if 'KeyboardInterrupt' in stderr.decode("utf-8"):
-                    success = None  # task cancelled
-                    break
+                if p.returncode != 0 or success is False:
+                    with open(os.path.join(self.task_dir, 'stdout.txt'), 'a') as o:
+                        o.write("Run no: %s, STDOUT at: %s\n" % (n+1, int(time())))
+                        o.write(stdout.decode("utf-8"))
+                    with open(os.path.join(self.task_dir, 'stderr.txt'), 'a') as e:
+                        e.write("Run no: %s, STDERR at: %s\n" % (n+1, int(time())))
+                        e.write(stderr.decode("utf-8"))
+                    if 'KeyboardInterrupt' in stderr.decode("utf-8"):
+                        success = None  # task cancelled
+                        break
+                    else:
+                        success = False  # task failed
                 else:
-                    success = False  # task failed
-            else:
-                break  # success
+                    break  # success
 
         time_end = int(time())
